requirements.txt
# Core Dependencies
openai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0

# HTTP Requests
requests>=2.31.0

# Logging and Utilities
# API
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
google-generativeai>=0.0.6
